Change all plotting logs from `info` to `debug` and simplify progress
This refactors how the progress of the plots is calculated, making it much
simplier. Before this change we would of had to keep track of the number of
times `percent_complete` was output, but now it simply updates the percent
complete after each amplicon is finished processing. Hopefully this will make
things easier to mantain even though it will be a little less "accurate" (not
sure how accurate the original implementation was...).