Refactor generating allele graph quilt using pandas
This refactor uses pandas methods to generate the structures needed to visualize
the allele plot as a graph. The previous method would add all of the nodes, and
then traverse the graph and merge duplicate editing events. This refactor
doesn't add any duplicates in the first place and therefore doesn't need to
merge anything and is more efficient.