Improve multiprocessing in Batch to utilize more processors
Previously, sub-crispresso runs are run on 1 process (with the rationale that parallelization of the alignment bit is the most important for batches). Now, with parallelized plotting, if more processes are provided than batches, each sub-crispresso run will be run with int(#processes/#batches) processes, thus maximizing the processor utilization for small batches.